Move useCurrentUser to sharedHooks/ and mock in ObsList test
